backend/src/requirements-minimal.txt
fastapi==0.115.4
uvicorn==0.32.0
langgraph==0.2.50
python-dotenv==1.0.0
supabase==2.3.4
groq==0.11.0